Place Order - Polished
Added order_number as GUID
Next move to Order History API get with pagination done and order DTO
Start FRONTEND

+++ DreamPlants.DataService.API/Controllers/AddressController.cs
@@ -4,6 +4,7 @@ using DreamPlants.DataService.API.Models.Generated;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.VisualStudio.Web.CodeGenerators.Mvc.Templates.BlazorIdentity.Pages.Manage;
+using Npgsql;
 using System.Security.Cryptography;
 using System.Text;
 
@@ -35,7 +36,7 @@ namespace DreamPlants.DataService.API.Controllers
           return Unauthorized(new { success = false, message = "Unauthorized" });
         // 3 - with user.UserId find user adresses
         List<AddressDTO> addressesDTO = await _context.Addresses
-            .Where(a => a.UserId == user.UserId)
+            .Where(a => a.UserId == user.UserId && !a.Deleted)
             .Select(a => new AddressDTO
             {
               AddressId = a.AddressId,
@@ -51,10 +52,10 @@ namespace DreamPlants.DataService.API.Controllers
 
         // 4 - if no adresses send ok with a message (Check if you need to send notfound) 
         if (addressesDTO == null || addressesDTO.Count == 0)
-          return Ok(new { success = true, message = "User doesnt have any adresses saved." });
+          return Ok(new { success = false, message = "User doesnt have any adresses saved." });
 
         // Last - Return Adresses
-        return Ok(new { success = true, message = "Address saved!", addressesDTO });
+        return Ok(new { success = true, message = "Address found!", addressesDTO });
       }
       catch (Exception ex)
       {
@@ -135,7 +136,7 @@ namespace DreamPlants.DataService.API.Controllers
         if (string.IsNullOrEmpty(token))
           return Unauthorized(new { success = false, message = "Unauthorized" });
 
-        User user = await _context.Users.FirstOrDefaultAsync(u => u.LoginToken == token);
+        var user = await _context.Users.FirstOrDefaultAsync(u => u.LoginToken == token);
         if (user == null)
           return Unauthorized(new { success = false, message = "Unauthorized" });
 
@@ -143,19 +144,32 @@ namespace DreamPlants.DataService.API.Controllers
         if (address == null)
           return Ok(new { success = false, message = "Address not found or does not belong to user." });
 
-        _context.Addresses.Remove(address);
-        await _context.SaveChangesAsync();
-
-        return Ok(new { success = true, message = "Address deleted!" });
+        // Better is a try catch - find out how to catch any or specific exceptions - fix later
+        bool isInUse = await _context.Orders.AnyAsync(o => o.AddressId == id);
+        if (isInUse)
+        {
+          // Soft delete
+          address.Deleted = true;
+          address.DeleteDate = DateTime.Now;
+          await _context.SaveChangesAsync();
+          return Ok(new { success = true, message = "Address deleted." });
+        }
+        else
+        {
+          // Hard delete
+          _context.Addresses.Remove(address);
+          await _context.SaveChangesAsync();
+          return Ok(new { success = true, message = "Address deleted!" });
+        }
       }
       catch (Exception ex)
       {
 #if DEBUG
         return StatusCode(500, new { success = false, message = ex.Message });
 #else
-        return StatusCode(500, new { success = false, message = "An error occurred." });
+    return StatusCode(500, new { success = false, message = "An error occurred." });
 #endif
       }
-    } // DelAddress
+    }    // DelAddress
   }
 }
